Added click on command option, warning, verbiage
If an option under Add a Step is not selected, the option for Dispatch to Nodes is not available. 

Also added a warning if the workflow step is not saved, the job can not be created. 
The notification of the missing step disappears two quickly in Chrome 72 when the create fails. Warning is needed ahead of time in the manual to prevent confusion.

Also changed verbiage from "check box" to "radio button" for the node selection run type.